distrib_systems_tasks/weeks/week-07/starter/requirements.txt
grpcio>=1.50.0
grpcio-tools>=1.50.0
protobuf>=4.21.0
pytest>=7.0.0